// src/app/model/CreatorProfile.ts
import { CreatorContent } from "./CreatorContent";

export class CreatorProfile {
    creatorName: string;
    creatorImage: string;
    coverImage: string;
    cantSeguidores: number;
    cantSubscriptores: number;
    contentDescription: string;
    biography: string;
    youtubeLink: string;
    plans: any[];
}